Total reorganization of proof annotations
Main idea: return the proof object along with the theorem, so that
printing the proof out looks less blatantly like it depends on
side-effects.

The annotation signature now provides anno_prove and anno_store_thm that
are like Q.prove and Q.store_thm except return proofs as well (as
generated by any anno_ tactics used in the tactic).

To print the proofs later, I store them in LoadableThyData with the
theory that proved the theorems. Thus the annotation signature also
provides the facility to turn a list of proofs into loadable theory data
and the script file does a write_data_update for each annotated theorem.

For proofs to be theory data, they need to be serializable. I use the
Pickle module from MLKit to convert proofs to and from strings, using
Parse.term_to_string and Parse.Term for the terms inside proofs.

Finally, since proofs can now be printed later, I moved the definition
of custom pretty printing things into a ppTools module, and the actual
adding of custom rules to the grammar and printing theorems/proofs to a
new theory ppTheory. This ppTheory is serves as the theory to give
the munger, and it's the only one with the TeXy rules loaded.